Fix UsbTiny programmer
As reported on https://github.com/arduino/Arduino/issues/8198 ,
`patch #8924: Enable TPI for usbtiny` broke the functionality of such programmer for non TPI targets.
The refactor included reworking `usbtiny_initialize`, which used to call `pgm->program_enable` once and, in case of failure, reset the programmer and call it again.
The new code considers the return code successful if >=0, so the first failure is not reported and the programmer is not being reset.
The attached patch restores the previous behaviour by checking if `pgm->program_enable() == 0`

Cross reference: https://savannah.nongnu.org/patch/index.php?9728